Make valid ES module
Hello @metonym ! Thank you for writing this library !

I'm using it and actually have a warning from Node.js : 

```sh
svelte-search doesn't appear to be written in CJS, but also doesn't appear to be a valid ES module (i.e. it doesn't have "type": "module" or an .mjs extension for the entry point). Please contact the package author to fix.
```

I think it's the only change needed for fixing it.